fixing preloads according to url
