+++ packages/cli/lib/lib/webpack/render-html-plugin.js
@@ -30,7 +30,8 @@ module.exports = function (config) {
 			manifest: config.manifest,
 			title: title || config.title || config.manifest.name || config.manifest.short_name || (config.pkg.name || '').replace(/^@[a-z]\//, '') || 'Preact App',
 			excludeAssets: [/(bundle|polyfills)(\..*)?\.js$/],
-			config,
+      config,
+      url,
 			ssr(params) {
 				Object.assign(params, { url });
 				return config.prerender ? prerender({ cwd, dest, src }, params) : '';
